backend/requirements.txt
fastapi
google-api-python-client
httpx
pandas
python-dotenv
uvicorn
scikit-learn
pydantic
jinja2
requests
python-multipart
openai

alembic
pytest
pytest-asyncio
python-jose

